+++ src/Math/Algebra/Hspray.hs
@@ -2446,7 +2446,8 @@ resultant var p q =
     det = detLaplace $ 
           sylvesterMatrix (sprayCoefficients p') (sprayCoefficients q')
 
--- | Subresultants of two sprays
+-- | Subresultants of two sprays (the /principal/ subresultants, while the 
+-- `polynomialSubresultants` function returns the polynomial subresultants)
 subresultants :: (Eq a, AlgRing.C a) 
   => Int     -- ^ indicator of the variable with respect to which the subresultants are desired (e.g. 1 for x)
   -> Spray a 
@@ -2533,32 +2534,14 @@ resultant' var sprayA sprayB
           h'  = exactDivisionBy (h^**^delta) (h ^*^ g'^**^delta)
           h'' = exactDivisionBy (h'^**^degp') (h' ^*^ ellq'^**^degp')
 
-{- sresMatrix :: (Eq a, AlgRing.C a) => Spray a -> Spray a -> Int -> Spray a
-sresMatrix p q i = if n == m && i == n 
-  then q
-  else detLaplace matrix
-  where
-    d = max (numberOfVariables p) (numberOfVariables q)
-    x = lone d
-    pcoeffs = reverse $ sprayCoefficients' d p
-    qcoeffs = reverse $ sprayCoefficients' d q
-    pcoeff k = if k < 0 then zeroSpray else pcoeffs !! k
-    qcoeff k = if k < 0 then zeroSpray else qcoeffs !! k
-    n = length pcoeffs - 1
-    m = length qcoeffs - 1
-    prow k = replicate k zeroSpray ++ 
-              [pcoeff j | j <- [n, n-1 .. 2*i - m + k + 2]] ++ 
-                [x^**^(m-i-1-k) ^*^ p]
-    prows = [prow k | k <- [0 .. m - i - 1]]
-    qrow k = replicate k zeroSpray ++ 
-              [qcoeff j | j <- [m, m-1 .. 2*i - n + k + 2]] ++
-                [x^**^(n-i-1-k) ^*^ q]
-    qrows = [qrow k | k <- [0 .. n - i - 1]]
-    matrix = fromLists (prows ++ qrows) 
- -}
-
+-- | Polynomial subresultants of two sprays (the `subresultants` function 
+-- computes the /principal/ subresultants)
 polynomialSubresultants :: 
-  (Eq a, AlgRing.C a) => Int -> Spray a -> Spray a -> [Spray a]
+    (Eq a, AlgRing.C a) 
+  => Int     -- ^ index of the variable with respect to which the subresultants will be computed (e.g. 2 for @y@)
+  -> Spray a 
+  -> Spray a 
+  -> [Spray a]
 polynomialSubresultants var p q 
   | var < 1 || var > d               
     = error "polynomialSubresultants: invalid variable index."